Added the ability to join and leave archived and public channels + the functionality surrounding these options.

+++ bot/src/button-commands/leave-channel.js
@@ -1,52 +1,54 @@
 import { getBot } from '../cache-bot.js'
-import { MessageActionRow, MessageButton } from 'discord.js'
 import { directMessageError } from '../utils/error-logging.js'
-import { getCommandSymbol } from '../repositories/guilds.js'
-import { getFormatedCommandChannels } from '../repositories/channels.js'
+import { getChannelType } from '../repositories/channels.js'
 
 export default async function (interaction) {
   const interactionChannel = getBot().channels.cache.get(
       interaction.customId.match(`(?!:)[0-9]+`)[0]
     ),
-    guild = interactionChannel.guild,
-    guildMember = guild.members.cache.get(interaction.user.id),
-    userOverwrite = interactionChannel.permissionOverwrites.cache.find(
+    guild = interaction.guild,
+    guildMember = guild.members.cache.get(interaction.user.id)
+
+  if (!interactionChannel) {
+    guildMember
+      .send(
+        `You tried leaving a channel that no longer exists, sorry for the trouble 🥺`
+      )
+      .catch(error => directMessageError(error, guildMember))
+
+    return
+  }
+
+  const channelType = await getChannelType(interactionChannel.id)
+
+  if (channelType === `private`) {
+    guildMember
+      .send(
+        `You tried leaving a channel that no longer exists, sorry for the trouble 🥺`
+      )
+      .catch(error => directMessageError(error, guildMember))
+
+    return
+  }
+
+  const userOverwrite = interactionChannel.permissionOverwrites.cache.find(
       permissionOverwrite => permissionOverwrite.id === guildMember.id
     ),
-    commandSymbol = await getCommandSymbol(guild.id),
-    commandChannels = await getFormatedCommandChannels(
-      guild.id,
-      `unrestricted`
-    ),
-    joinButtonRow = new MessageActionRow().addComponents(
-      new MessageButton()
-        .setCustomId(`!join-channel: ${interactionChannel.id}`)
-        .setLabel(`Join ${interactionChannel.name}`)
-        .setStyle('SUCCESS')
-    )
+    individualPermissions = userOverwrite
+      ? userOverwrite.allow.serialize()
+      : null
 
-  if (userOverwrite) {
-    userOverwrite.delete()
-
-    guildMember
-      .send({
-        content: `
-          You have been removed from <#${interactionChannel.id}> in the ${guild.name} server 👋\
-          \nIf you left by accident press the button below, or use the \`${commandSymbol}join\` command (ex: \`${commandSymbol}join ${interactionChannel.name}\`) to re-join.\
-          \nThe \`${commandSymbol}join\` command works in these channels: ${commandChannels}
-        `,
-        components: [joinButtonRow],
+  if ([`archived`, `joinable`].includes(channelType)) {
+    if (userOverwrite) userOverwrite.delete()
+  } else if (channelType === `public`) {
+    if (!userOverwrite) {
+      interactionChannel.permissionOverwrites.create(guildMember.id, {
+        VIEW_CHANNEL: false,
       })
-      .catch(error => directMessageError(error, guildMember))
-  } else
-    guildMember
-      .send({
-        content: `
-          You tried to leave a channel you aren't a part of, <#${interactionChannel.id}> in the ${guild.name} server 🤔\
-          \nIf you'd like to join press the button below, or use the \`${commandSymbol}join\` command (ex: \`${commandSymbol}join ${interactionChannel.name}\`).\
-          \nThe \`${commandSymbol}join\` command works in these channels: ${commandChannels}
-        `,
-        components: [joinButtonRow],
+    } else if (individualPermissions?.VIEW_CHANNEL) {
+      interactionChannel.permissionOverwrites.edit(guildMember.id, {
+        VIEW_CHANNEL: false,
       })
-      .catch(error => directMessageError(error, guildMember))
+    }
+  }
 }
